original IDs need to be bigint

+++ db/migrate/20211210025522_create_wunderlists.rb
@@ -3,7 +3,7 @@ class CreateWunderlists < ActiveRecord::Migration[6.1]
     create_table :wunderlists do |t|
       t.belongs_to :wunderfolder, null: false, foreign_key: true
 
-      t.integer  :original_id
+      t.bigint   :original_id
       t.string   :title
       t.string   :directory_path
       t.string   :text_file_path
